Structural cleanups
Freeware Atari800 smoothed font that may work nicely (it's hard to find
a free font with an explicit license :/ ).
List of stages of the game; each of these should probably be a subclass
of the UI/Logic or some such contortion.
Game state object strimmed down to just be data storage. Various
accessors and mutators on other involved classes.
Weakening 'Resource' type to include 'None', so that Tiles can be
owned but not equipped without adding another boolean and making the
code that has to use this even uglier.
Splitting main into the actual main, which deals with argv uglyness,
and a 'realmain' which does the init/run/shutdown of the game.
Adding the shell of the 'sprite' UI, and main code to just about make a
UI go.
This involves a rather gratuitous factory, really, but it avoids the
need to ever expose implementation details behind the interface.

git-svn-id: svn+ssh://private/mule@1466 51cd4cf5-f2fb-0310-83f2-91b3fd76ea51