Remove Dummies, initialize game state
Sprite UI now initialises SDL_mixer and SDL_ttf; loads a font for the
latter.
Add fullscreen toggling on Escape and now make it the default.
Remove the DummyPlayer and (disable) the DummyController, as the
change from references to pointers makes them unncessary (and
incorrect).
Change Terrain getSize and getCity to be pairs of functions instead of
functions which return pairs, as pairs are really clunky for this and
there's no recomputation to avoid when just returning members.
Game constructor now initialises all of its state (including Terrain
and Players) based on the PlayerSetups and Difficulty, including
terrain generation.

git-svn-id: svn+ssh://private/mule@1467 51cd4cf5-f2fb-0310-83f2-91b3fd76ea51