Initial work on graphics
Made some developer-time graphics, but not checking in yet.
Hashes for sounds and textures, and methods to load such, in the Sprite
UI.
GameState header, for the overall state of the game as a program (vs
Game, which is the state of an abstract game MULE).
Add SDL_image to Makefile libraries, as it's used for texture-loading.

git-svn-id: svn+ssh://private/mule@1468 51cd4cf5-f2fb-0310-83f2-91b3fd76ea51