Add more implementation docs/notes
Add a doc/ folder with a graphviz dot diagram of state transitions in
it.
Updated state list in notes with the per-state shared logic/UI data
required.
Turns out M.U.L.E. has a wampus, not a wumpus. Misread that /how/ many
years?
Change to faster text rendering routine out of general contraryness;
mostly because it allows easy pallette shenanigans.
Note about how "UI" is a bit of a misnomer when the logic is doing all
the input and behaviour.
Poke a few gamelogic-related bits of main.
Make GameLogic destructor public so that it can be deleted by main. Fix
comment to be doxygenable.

git-svn-id: svn+ssh://private/mule@1544 51cd4cf5-f2fb-0310-83f2-91b3fd76ea51