New pointer graphics and code
New pointer graphics as SVG, which have a larger player-colour area,
less overall precision (as this is not an exact clicking game), and
directional cues (because that is important for the auction, etc.).
Fix g++ pedantry about return values.
New class UserInterfaceSpritePointer, which is a special Sprite
subclass for pointers, performing the necessary channel remapping,
rotation, and transparent pixmap switching to act as a new-style
pointer while still rendering with all the other sprites.
The new pointer class is not yet hooked in or tested.
Makefile autogenerates pointer PNGs from pointer SVGs using Inkscape.

git-svn-id: svn+ssh://private/mule@1624 51cd4cf5-f2fb-0310-83f2-91b3fd76ea51