Fix use of IMG_Init(), controllers
Make IMG_Init() conditional on latest API version, as apparently it's a
new addition. That would be how I missed it. A new addition in a minor
version. Sigh.

Change arrows controller to use right shift, not right control, as this
works better on laptops.

Have controller agree with main on joystick ball and hat events.
(Apparently the Logitech gamepads send these for the digital pad when
in analogue mode. THE MORE YOU KNOW.)

git-svn-id: svn+ssh://private/mule@1716 51cd4cf5-f2fb-0310-83f2-91b3fd76ea51