C++11; Fix segfault from incorrect SDL shutdown
If you quit from the titlescreen you could often trigger a segfault,
since the JoystickController was using SDL calls to release joystick
handles after SDL_Quit() had been called. Use a unique_ptr to narrow
down the lifespan of the ControlManager so it destructs before the
call to SDL_Quit().

Requires upgrading to C++11, which is a good idea anyway.